Move SessionStream from `imap` to `net`
