+++ src/imap/session.rs
@@ -1,16 +1,10 @@
 use std::ops::{Deref, DerefMut};
-use std::pin::Pin;
-use std::time::Duration;
 
 use async_imap::types::Mailbox;
 use async_imap::Session as ImapSession;
-use async_native_tls::TlsStream;
-use fast_socks5::client::Socks5Stream;
-use tokio::io::BufWriter;
-use tokio::net::TcpStream;
-use tokio_io_timeout::TimeoutStream;
 
-use super::capabilities::Capabilities;
+use crate::imap::capabilities::Capabilities;
+use crate::net::session::SessionStream;
 
 #[derive(Debug)]
 pub(crate) struct Session {
@@ -25,39 +19,6 @@ pub(crate) struct Session {
     pub selected_mailbox: Option<Mailbox>,
 
     pub selected_folder_needs_expunge: bool,
-}
-
-pub(crate) trait SessionStream:
-    tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin + Send + Sync + std::fmt::Debug
-{
-    /// Change the read timeout on the session stream.
-    fn set_read_timeout(&mut self, timeout: Option<Duration>);
-}
-
-impl SessionStream for Box<dyn SessionStream> {
-    fn set_read_timeout(&mut self, timeout: Option<Duration>) {
-        self.as_mut().set_read_timeout(timeout);
-    }
-}
-impl<T: SessionStream> SessionStream for TlsStream<T> {
-    fn set_read_timeout(&mut self, timeout: Option<Duration>) {
-        self.get_mut().set_read_timeout(timeout);
-    }
-}
-impl<T: SessionStream> SessionStream for BufWriter<T> {
-    fn set_read_timeout(&mut self, timeout: Option<Duration>) {
-        self.get_mut().set_read_timeout(timeout);
-    }
-}
-impl SessionStream for Pin<Box<TimeoutStream<TcpStream>>> {
-    fn set_read_timeout(&mut self, timeout: Option<Duration>) {
-        self.as_mut().set_read_timeout_pinned(timeout);
-    }
-}
-impl<T: SessionStream> SessionStream for Socks5Stream<T> {
-    fn set_read_timeout(&mut self, timeout: Option<Duration>) {
-        self.get_socket_mut().set_read_timeout(timeout)
-    }
 }
 
 impl Deref for Session {
